test: pass props but failed really need help
